+++ CHANGES.md
@@ -18,6 +18,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - build: add check for isort command
 
+### Meta
+
+- test: add coverage and pytest-coverage. remove tox
+
 ## [0.1.1] - 2021-12-10
 
 ### Fixed