+++ CHANGES.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-12-14
+
 ### Added
 
 - BREAKING: added `RoomServiceClient` wrapper class
@@ -42,9 +44,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Initial release
 - [LiveKit protocol v0.11.0](https://github.com/livekit/protocol/releases/tag/v0.11.0)
 
-[Unreleased]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.2.0...HEAD
 
-[0.2.0]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.2.0...v0.1.1
+[0.2.0]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.1.1...v0.2.0
 
 [0.1.1]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.1.0...v0.1.1